2.5.4

+++ app/src/main/java/eu/kanade/tachiyomi/ui/reader/ReaderPresenter.kt
@@ -19,7 +19,6 @@ import eu.kanade.tachiyomi.data.download.DownloadManager
 import eu.kanade.tachiyomi.data.notification.NotificationReceiver
 import eu.kanade.tachiyomi.data.notification.Notifications
 import eu.kanade.tachiyomi.data.preference.PreferencesHelper
-import eu.kanade.tachiyomi.data.preference.getOrDefault
 import eu.kanade.tachiyomi.data.track.TrackManager
 import eu.kanade.tachiyomi.jobs.tracking.DelayedTrackingUpdateJob
 import eu.kanade.tachiyomi.source.SourceManager
@@ -33,7 +32,10 @@ import eu.kanade.tachiyomi.ui.reader.loader.ChapterLoader
 import eu.kanade.tachiyomi.ui.reader.model.ReaderChapter
 import eu.kanade.tachiyomi.ui.reader.model.ReaderPage
 import eu.kanade.tachiyomi.ui.reader.model.ViewerChapters
+import eu.kanade.tachiyomi.ui.reader.settings.OrientationType
+import eu.kanade.tachiyomi.ui.reader.settings.ReadingModeType
 import eu.kanade.tachiyomi.util.chapter.ChapterFilter
+import eu.kanade.tachiyomi.util.chapter.ChapterSort
 import eu.kanade.tachiyomi.util.chapter.syncChaptersWithSource
 import eu.kanade.tachiyomi.util.storage.DiskUtil
 import eu.kanade.tachiyomi.util.system.ImageUtil
@@ -118,9 +120,11 @@ class ReaderPresenter(
         val chaptersForReader =
             chapterFilter.filterChaptersForReader(dbChapters, manga, selectedChapter)
 
-        when (manga.sorting) {
-            Manga.SORTING_SOURCE -> ChapterLoadBySource().get(chaptersForReader)
-            Manga.SORTING_NUMBER -> ChapterLoadByNumber().get(chaptersForReader, selectedChapter)
+        when (manga.chapterOrder(preferences)) {
+            Manga.CHAPTER_SORTING_SOURCE -> ChapterLoadBySource().get(chaptersForReader)
+            Manga.CHAPTER_SORTING_NUMBER -> ChapterLoadByNumber().get(chaptersForReader,
+                selectedChapter)
+            Manga.CHAPTER_SORTING_UPLOAD_DATE -> ChapterLoadByDate().get(chaptersForReader)
             else -> error("Unknown sorting method")
         }.map(::ReaderChapter)
     }
@@ -215,49 +219,20 @@ class ReaderPresenter(
     suspend fun getChapters(): List<ReaderChapterItem> {
         val manga = manga ?: return emptyList()
         chapterItems = withContext(Dispatchers.IO) {
+            val chapterSort = ChapterSort(manga, chapterFilter, preferences)
             val dbChapters = db.getChapters(manga).executeAsBlocking()
-            val currentChapter = getCurrentChapter()?.chapter?.id
-            val list =
-                chapterFilter.filterChaptersForReader(
-                    dbChapters,
+            chapterSort.getChaptersSorted(dbChapters,
+                filterForReader = true,
+                currentChapter = getCurrentChapter()?.chapter).map {
+                ReaderChapterItem(
+                    it,
                     manga,
-                    getCurrentChapter()?.chapter
+                    it.id == getCurrentChapter()?.chapter?.id ?: chapterId
                 )
-                    .sortedBy {
-                        when (manga.sorting) {
-                            Manga.SORTING_NUMBER -> it.chapter_number
-                            else -> it.source_order.toFloat()
-                        }
-                    }.map {
-                        ReaderChapterItem(
-                            it,
-                            manga,
-                            it.id == currentChapter ?: chapterId
-                        )
-                    }
-            if (!manga.sortDescending(preferences.chaptersDescAsDefault().getOrDefault())) {
-                list.reversed()
-            } else {
-                list
             }
         }
 
         return chapterItems
-    }
-
-    /**
-     * Removes all filters and requests an UI update.
-     */
-    fun setFilters(read: Boolean, unread: Boolean, downloaded: Boolean, bookmarked: Boolean) {
-        val manga = manga ?: return
-        manga.readFilter = when {
-            read -> Manga.SHOW_READ
-            unread -> Manga.SHOW_UNREAD
-            else -> Manga.SHOW_ALL
-        }
-        manga.downloadedFilter = if (downloaded) Manga.SHOW_DOWNLOADED else Manga.SHOW_ALL
-        manga.bookmarkedFilter = if (bookmarked) Manga.SHOW_BOOKMARKED else Manga.SHOW_ALL
-        db.updateFlags(manga).executeAsBlocking()
     }
 
     /**
@@ -477,11 +452,9 @@ class ReaderPresenter(
         val selectedChapter = page.chapter
 
         // Save last page read and mark as read if needed
-        selectedChapter.chapter.last_page_read = page.index
-        selectedChapter.chapter.pages_left =
-            (selectedChapter.pages?.size ?: page.index) - page.index
+
         val shouldTrack =
-            !preferences.incognitoMode().get() || hasTrackers || preferences.readingSync()
+            preferences.incognitoMode().get().not() || hasTrackers || preferences.readingSync()
         if (shouldTrack &&
             // For double pages, check if the second to last page is doubled up
             (
@@ -489,10 +462,15 @@ class ReaderPresenter(
                     (hasExtraPage && selectedChapter.pages?.lastIndex?.minus(1) == page.index)
                 )
         ) {
-            selectedChapter.chapter.read = true
-            updateTrackChapterRead(selectedChapter)
-            updateReadingStatus(selectedChapter)
-            deleteChapterIfNeeded(selectedChapter)
+            if (preferences.incognitoMode().get().not()) {
+                selectedChapter.chapter.last_page_read = page.index
+                selectedChapter.chapter.pages_left =
+                    (selectedChapter.pages?.size ?: page.index) - page.index
+                selectedChapter.chapter.read = true
+                updateTrackChapterRead(selectedChapter)
+                updateReadingStatus(selectedChapter)
+                deleteChapterIfNeeded(selectedChapter)
+            }
         }
 
         if (selectedChapter != currentChapters.currChapter) {
@@ -568,24 +546,26 @@ class ReaderPresenter(
     /**
      * Returns the viewer position used by this manga or the default one.
      */
-    fun getMangaViewer(): Int {
-        val default = preferences.defaultViewer()
+    fun getMangaReadingMode(): Int {
+        val default = preferences.defaultReadingMode()
         val manga = manga ?: return default
         val readerType = manga.defaultReaderType()
-        if (manga.viewer == -1 ||
+        if (manga.viewer_flags == -1 ||
             // Force webtoon mode
-            (manga.isLongStrip() && readerType != manga.viewer)
+            (manga.isLongStrip() && readerType != manga.readingModeType)
         ) {
             val cantSwitchToLTR =
-                (readerType == ReaderActivity.LEFT_TO_RIGHT && default != ReaderActivity.RIGHT_TO_LEFT)
-            manga.viewer = if (cantSwitchToLTR) 0 else readerType
-            db.updateMangaViewer(manga).asRxObservable().subscribe()
+                (
+                    readerType == ReadingModeType.LEFT_TO_RIGHT.flagValue &&
+                        default != ReadingModeType.RIGHT_TO_LEFT.flagValue
+                    )
+            manga.readingModeType = if (cantSwitchToLTR) 0 else readerType
+            db.updateViewerFlags(manga).asRxObservable().subscribe()
         }
-
-        val viewer = if (manga.viewer == 0) preferences.defaultViewer() else manga.viewer
+        val viewer = if (manga.readingModeType == 0) default else manga.readingModeType
 
         return when {
-            !manga.isLongStrip() && viewer == ReaderActivity.WEBTOON -> ReaderActivity.VERTICAL_PLUS
+            !manga.isLongStrip() && viewer == ReadingModeType.WEBTOON.flagValue -> ReadingModeType.CONTINUOUS_VERTICAL.flagValue
             else -> viewer
         }
     }
@@ -593,10 +573,10 @@ class ReaderPresenter(
     /**
      * Updates the viewer position for the open manga.
      */
-    fun setMangaViewer(viewer: Int) {
+    fun setMangaReadingMode(readingModeType: Int) {
         val manga = manga ?: return
-        manga.viewer = viewer
-        db.updateMangaViewer(manga).executeAsBlocking()
+        manga.readingModeType = readingModeType
+        db.updateViewerFlags(manga).executeAsBlocking()
 
         Observable.timer(250, TimeUnit.MILLISECONDS, AndroidSchedulers.mainThread())
             .subscribeFirst({ view, _ ->
@@ -614,9 +594,44 @@ class ReaderPresenter(
     }
 
     /**
+     * Returns the orientation type used by this manga or the default one.
+     */
+    fun getMangaOrientationType(): Int {
+        val default = preferences.defaultOrientationType().get()
+        return when (manga?.orientationType) {
+            OrientationType.DEFAULT.flagValue -> default
+            else -> manga?.orientationType ?: default
+        }
+    }
+
+    /**
+     * Updates the orientation type for the open manga.
+     */
+    fun setMangaOrientationType(rotationType: Int) {
+        val manga = manga ?: return
+        manga.orientationType = rotationType
+        db.updateViewerFlags(manga).executeAsBlocking()
+
+        XLog.i("Manga orientation is ${manga.orientationType}")
+
+        Observable.timer(250, TimeUnit.MILLISECONDS, AndroidSchedulers.mainThread())
+            .subscribeFirst({ view, _ ->
+                val currChapters = viewerChaptersRelay.value
+                if (currChapters != null) {
+                    view.setOrientation(getMangaOrientationType())
+                }
+            })
+    }
+
+    /**
      * Saves the image of this [page] in the given [directory] and returns the file location.
      */
-    private fun saveImage(page: ReaderPage, directory: File, manga: Manga): File {
+    private fun saveImage(
+        page: ReaderPage,
+        directory: File,
+        manga: Manga,
+        prefix: String = "",
+    ): File {
         val stream = page.stream!!
         val type = ImageUtil.findImageType(stream) ?: throw Exception("Not an image")
 
@@ -627,7 +642,7 @@ class ReaderPresenter(
         // create chapter name so its always sorted correctly  max character is 75
         val pageName = parseChapterName(chapter.name, page.number.toString(), chapter.scanlator)
         // take only 150 characters so this file maxes at 225
-        val trimmedTitle = manga.title.take(150)
+        val trimmedTitle = (prefix + manga.title).take(150)
 
         // Build destination file
         val filename =
@@ -802,7 +817,7 @@ class ReaderPresenter(
         val destDir = File(context.cacheDir, "shared_image")
 
         Observable.fromCallable { destDir.deleteRecursively() } // Keep only the last shared file
-            .map { saveImage(page, destDir, manga) }
+            .map { saveImage(page, destDir, manga, "SPOILER_") }
             .subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread())
             .subscribeFirst(
